Refactor seventh lab

+++ src/main/java/com/example/oop6/models/shapes/Circle.java
@@ -1,9 +1,9 @@
 package com.example.oop6.models.shapes;
 
 import javafx.scene.canvas.GraphicsContext;
-
+/* Эллипс */
 public class Circle extends Shape {
-    public Circle(int width, int height) {
+    public Circle(double width, double height) {
         super(width, height);
     }
