+++ src/main/java/com/example/oop6/models/shapes/Triangle.java
@@ -2,8 +2,9 @@ package com.example.oop6.models.shapes;
 
 import javafx.scene.canvas.GraphicsContext;
 
+/* Треугольник */
 public class Triangle extends Shape {
-    public Triangle(int width, int height) {
+    public Triangle(double width, double height) {
         super(width, height);
     }
